moved most things to machine level
