+++ spynnaker/pyNN/external_devices_models/machine_munich_motor_device.py
@@ -14,35 +14,88 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from spinn_utilities.overrides import overrides
+from pacman.model.constraints.key_allocator_constraints import (
+    FixedMaskConstraint)
+from pacman.executor.injection_decorator import inject_items
 from pacman.model.graphs.common import Slice
 from pacman.model.graphs.machine import MachineVertex
 from pacman.model.resources import (
     ConstantSDRAM, CPUCyclesPerTickResource, DTCMResource, ResourceContainer)
 from spinn_front_end_common.abstract_models import (
     AbstractHasAssociatedBinary)
+from spinn_front_end_common.abstract_models import (
+    AbstractGeneratesDataSpecification,
+    AbstractProvidesOutgoingPartitionConstraints)
+from spinn_front_end_common.abstract_models.impl import (
+    ProvidesKeyToAtomMappingImpl)
 from spinn_front_end_common.interface.provenance import \
     ProvidesProvenanceDataFromMachineImpl
+from spinn_front_end_common.interface.simulation import simulation_utilities
 from spinn_front_end_common.utilities.constants import (
     SYSTEM_BYTES_REQUIREMENT, SIMULATION_N_BYTES, BYTES_PER_WORD)
-from spinn_front_end_common.utilities.utility_objs import ExecutableType
+from spinn_front_end_common.utilities.utility_objs import (
+    ExecutableType, ProvenanceDataItem)
+from spynnaker.pyNN.exceptions import SpynnakerException
 
 
-class MachineMunichMotorDevice(MachineVertex, AbstractHasAssociatedBinary):
+class MachineMunichMotorDevice(
+    MachineVertex, AbstractGeneratesDataSpecification,
+    AbstractHasAssociatedBinary, AbstractProvidesOutgoingPartitionConstraints,
+    ProvidesKeyToAtomMappingImpl, ProvidesProvenanceDataFromMachineImpl):
     """ An Omnibot motor control device. This has a real vertex and an \
         external device vertex.
     """
-    _slots__ = []
+    __slots__ = [
+        "__continue_if_not_different",
+        "__delay_time",
+        "__delta_threshold",
+        "__sample_time",
+        "__speed",
+        "__update_time"]
+
+    MOTOR_PARTITION_ID = "MOTOR"
 
     N_ATOMS = 6
+
+    SYSTEM_REGION = 0
+    PARAMS_REGION = 1
+    PROVENANCE_REGION = 2
 
     PROVENANCE_ELEMENTS = 1
 
     PARAMS_SIZE = 7 * BYTES_PER_WORD
 
-    def __init__(self, label=None, constraints=None, app_vertex=None):
+    INPUT_BUFFER_FULL_NAME = "Times_the_input_buffer_lost_packets"
+    INPUT_BUFFER_FULL_MESSAGE = (
+        "The input buffer for {} on {}, {}, {} lost packets on {} "
+        "occasions. This is often a sign that the system is running "
+        "too quickly for the number of neurons per core.  Please "
+        "increase the timer_tic or time_scale_factor or decrease the "
+        "number of neurons per core.")
+    def __init__(
+            self, speed, sample_time, update_time, delay_time,
+            delta_threshold, continue_if_not_different,
+            label=None, constraints=None, app_vertex=None):
+        """
+        :param int speed:
+        :param int sample_time:
+        :param int update_time:
+        :param int delay_time:
+        :param int delta_threshold:
+        :param bool continue_if_not_different:
+        :param str label:
+        :param constraints:
+        :param app_vertex:
+        """
         super(MachineMunichMotorDevice, self).__init__(
             label=label, constraints=constraints, app_vertex=app_vertex,
             vertex_slice=Slice(0, self.N_ATOMS-1))
+        self.__speed = speed
+        self.__sample_time = sample_time
+        self.__update_time = update_time
+        self.__delay_time = delay_time
+        self.__delta_threshold = delta_threshold
+        self.__continue_if_not_different = bool(continue_if_not_different)
 
     @property
     @overrides(MachineVertex.resources_required)
@@ -50,7 +103,7 @@ class MachineMunichMotorDevice(MachineVertex, AbstractHasAssociatedBinary):
         return ResourceContainer(
             sdram=ConstantSDRAM(
                 SYSTEM_BYTES_REQUIREMENT + self.PARAMS_SIZE +
-                ProvidesProvenanceDataFromMachineImpl.get_provenance_data_size(self.PROVENANCE_ELEMENTS)),
+                self.get_provenance_data_size(self.PROVENANCE_ELEMENTS)),
             dtcm=DTCMResource(0), cpu_cycles=CPUCyclesPerTickResource(0))
 
     @overrides(AbstractHasAssociatedBinary.get_binary_file_name)
@@ -60,3 +113,115 @@ class MachineMunichMotorDevice(MachineVertex, AbstractHasAssociatedBinary):
     @overrides(AbstractHasAssociatedBinary.get_binary_start_type)
     def get_binary_start_type(self):
         return ExecutableType.USES_SIMULATION_INTERFACE
+
+    @property
+    @overrides(ProvidesProvenanceDataFromMachineImpl._provenance_region_id)
+    def _provenance_region_id(self):
+        return self.PROVENANCE_REGION
+
+    @property
+    @overrides(ProvidesProvenanceDataFromMachineImpl._n_additional_data_items)
+    def _n_additional_data_items(self):
+        return self.PROVENANCE_ELEMENTS
+
+    @overrides(ProvidesProvenanceDataFromMachineImpl.
+               get_provenance_data_from_machine)
+    def get_provenance_data_from_machine(self, transceiver, placement):
+        # get prov data
+        provenance_data = self._read_provenance_data(transceiver, placement)
+        # get system level prov
+        provenance_items = self._read_basic_provenance_items(
+            provenance_data, placement)
+        # get left over prov
+        provenance_data = self._get_remaining_provenance_data_items(
+            provenance_data)
+        # stuff for making prov data items
+        label, x, y, p, names = self._get_placement_details(placement)
+
+        # get the only app level prov item
+        n_buffer_overflows = provenance_data[0]
+
+        # build it
+        provenance_items.append(ProvenanceDataItem(
+            self._add_name(names, self.INPUT_BUFFER_FULL_NAME),
+            n_buffer_overflows, report=n_buffer_overflows > 0,
+            message=self.INPUT_BUFFER_FULL_MESSAGE.format(
+                label, x, y, p, n_buffer_overflows)))
+        return provenance_items
+
+    @overrides(AbstractProvidesOutgoingPartitionConstraints.
+               get_outgoing_partition_constraints)
+    def get_outgoing_partition_constraints(self, partition):
+
+        # Any key to the device will work, as long as it doesn't set the
+        # management bit.  We also need enough for the configuration bits
+        # and the management bit anyway
+        return list([FixedMaskConstraint(0xFFFFF800)])
+
+    @inject_items({
+        "routing_info": "MemoryRoutingInfos",
+        "machine_time_step": "MachineTimeStep",
+        "time_scale_factor": "TimeScaleFactor"
+    })
+    @overrides(
+        AbstractGeneratesDataSpecification.generate_data_specification,
+        additional_arguments={
+            "routing_info", "machine_time_step", "time_scale_factor"
+        })
+    def generate_data_specification(
+            self, spec, placement, routing_info,
+            machine_time_step, time_scale_factor):
+        # pylint: disable=too-many-arguments, arguments-differ
+
+        # reserve regions
+        self.reserve_memory_regions(spec)
+
+        # Write the setup region
+        spec.comment("\n*** Spec for robot motor control ***\n\n")
+
+        # handle simulation data
+        spec.switch_write_focus(self.SYSTEM_REGION)
+        spec.write_array(simulation_utilities.get_simulation_header_array(
+            placement.vertex.get_binary_file_name(), machine_time_step,
+            time_scale_factor))
+
+        # Get the key
+        edge_key = routing_info.get_first_key_from_pre_vertex(
+            placement.vertex, self.MOTOR_PARTITION_ID)
+        if edge_key is None:
+            raise SpynnakerException(
+                "This motor should have one outgoing edge to the robot")
+
+        # write params to memory
+        spec.switch_write_focus(region=self.PARAMS_REGION)
+        spec.write_value(data=edge_key)
+        spec.write_value(data=self.__speed)
+        spec.write_value(data=self.__sample_time)
+        spec.write_value(data=self.__update_time)
+        spec.write_value(data=self.__delay_time)
+        spec.write_value(data=self.__delta_threshold)
+        spec.write_value(data=int(self.__continue_if_not_different))
+
+        # End-of-Spec:
+        spec.end_specification()
+
+    def reserve_memory_regions(self, spec):
+        """ Reserve SDRAM space for memory areas:
+
+        #. Area for information on what data to record
+        #. area for start commands
+        #. area for end commands
+
+        :param spec: The data specification to write to
+        :type spec: ~data_specification.DataSpecificationGenerator
+        """
+        spec.comment("\nReserving memory space for data regions:\n\n")
+
+        # Reserve memory:
+        spec.reserve_memory_region(
+            self.SYSTEM_REGION, SIMULATION_N_BYTES, label='setup')
+
+        spec.reserve_memory_region(
+            self.PARAMS_REGION, self.PARAMS_SIZE, label='params')
+
+        self.reserve_provenance_data_region(spec)
